# temp_openapi.json
{"openapi":"3.1.0","info":{"title":"CodeRAG API","description":"Chat with your GitHub repositories using RAG and code-aware context.","version":"0.1.0"},"paths":{"/api/chat":{"post":{"tags":["chat"],"summary":"Chat","description":"Synchronous chat: answer a question about a repository.\n\nBody:\n    {\n        \"repo_url\": \"https://github.com/user/repo\",\n        \"question\": \"How does authorization work?\",\n        \"max_chunks\": 15,\n        \"stream\": false\n    }","operationId":"chat_api_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/chat/stream":{"post":{"tags":["chat"],"summary":"Chat Stream","description":"Streaming chat via Server-Sent Events.\n\nReturns SSE events:\n  - event: \"token\", data: \"text chunk\"\n  - event: \"sources\", data: JSON with sources array\n  - event: \"done\", data: \"true\"","operationId":"chat_stream_api_chat_stream_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/repo/clone":{"post":{"tags":["repo"],"summary":"Clone Repo","description":"Clone or update a repository and start indexing.\n\nBody:\n    {\n        \"repo_url\": \"https://github.com/user/repo\",\n        \"branch\": \"main\",\n        \"force_reindex\": false\n    }","operationId":"clone_repo_api_repo_clone_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RepoCloneRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RepoStatusResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/repo/status/{repo_url}":{"get":{"tags":["repo"],"summary":"Get Repo Status","description":"Get the current indexing status of a repository.\n\nReturns:\n    RepoStatusResponse with status and stats.","operationId":"get_repo_status_api_repo_status__repo_url__get","parameters":[{"name":"repo_url","in":"path","required":true,"schema":{"type":"string","title":"Repo Url"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RepoStatusResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/repo/list":{"get":{"tags":["repo"],"summary":"List Repos","description":"List all cloned repositories.\n\nReturns:\n    List of repo names.","operationId":"list_repos_api_repo_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/search":{"post":{"tags":["search"],"summary":"Search","description":"Perform a raw vector search in Qdrant.\n\nBody:\n    {\n        \"repo_url\": \"https://github.com/user/repo\",\n        \"query\": \"authorization function\",\n        \"top_k\": 10,\n        \"language\": \"python\",\n        \"symbol_type\": \"function\"\n    }","operationId":"search_api_search_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/SearchResult"},"type":"array","title":"Response Search Api Search Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/health":{"get":{"tags":["health"],"summary":"Health Check","description":"Check health of all services.\n\nReturns:\n    {\n        \"status\": \"healthy|unhealthy\",\n        \"services\": {\n            \"vector_store\": \"healthy|unhealthy\",\n            \"llm\": \"healthy|unhealthy\"\n        }\n    }","operationId":"health_check_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ChatRequest":{"properties":{"repo_url":{"type":"string","title":"Repo Url"},"question":{"type":"string","title":"Question"},"max_chunks":{"type":"integer","title":"Max Chunks","default":15},"stream":{"type":"boolean","title":"Stream","default":false}},"type":"object","required":["repo_url","question"],"title":"ChatRequest"},"ChatResponse":{"properties":{"answer":{"type":"string","title":"Answer"},"sources":{"items":{"$ref":"#/components/schemas/SourceReference"},"type":"array","title":"Sources"},"repo_name":{"type":"string","title":"Repo Name"},"question":{"type":"string","title":"Question"}},"type":"object","required":["answer","sources","repo_name","question"],"title":"ChatResponse"},"ChunkMetadata":{"properties":{"symbol_id":{"type":"string","title":"Symbol Id"},"symbol_name":{"type":"string","title":"Symbol Name"},"symbol_type":{"$ref":"#/components/schemas/SymbolType"},"file_path":{"type":"string","title":"File Path"},"start_line":{"type":"integer","title":"Start Line"},"end_line":{"type":"integer","title":"End Line"},"language":{"type":"string","title":"Language"},"signature":{"type":"string","title":"Signature","default":""},"parent_class":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Parent Class"},"repo_name":{"type":"string","title":"Repo Name"},"repo_url":{"type":"string","title":"Repo Url","default":""},"branch":{"type":"string","title":"Branch","default":"main"},"dependencies":{"items":{"type":"string"},"type":"array","title":"Dependencies"},"chunk_type":{"type":"string","title":"Chunk Type","default":"code"}},"type":"object","required":["symbol_id","symbol_name","symbol_type","file_path","start_line","end_line","language","repo_name"],"title":"ChunkMetadata","description":"Payload stored alongside each embedding vector in Qdrant."},"CodeChunk":{"properties":{"id":{"type":"string","title":"Id"},"symbol_id":{"type":"string","title":"Symbol Id"},"text":{"type":"string","title":"Text"},"metadata":{"$ref":"#/components/schemas/ChunkMetadata"},"embedding":{"anyOf":[{"items":{"type":"number"},"type":"array"},{"type":"null"}],"title":"Embedding"}},"type":"object","required":["id","symbol_id","text","metadata"],"title":"CodeChunk","description":"A chunk ready for embedding and storing in Qdrant."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"},"services":{"type":"object","title":"Services"}},"type":"object","required":["status","services"],"title":"HealthResponse"},"RepoCloneRequest":{"properties":{"repo_url":{"type":"string","title":"Repo Url","description":"GitHub repo URL, e.g. https://github.com/user/repo.git"},"branch":{"type":"string","title":"Branch","default":"main"},"force_reindex":{"type":"boolean","title":"Force Reindex","default":false}},"type":"object","required":["repo_url"],"title":"RepoCloneRequest"},"RepoStatus":{"type":"string","enum":["cloning","indexing","ready","error","not_found"],"title":"RepoStatus"},"RepoStatusResponse":{"properties":{"repo_url":{"type":"string","title":"Repo Url"},"repo_name":{"type":"string","title":"Repo Name"},"status":{"$ref":"#/components/schemas/RepoStatus"},"files_indexed":{"type":"integer","title":"Files Indexed","default":0},"symbols_indexed":{"type":"integer","title":"Symbols Indexed","default":0},"last_indexed":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Indexed"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"}},"type":"object","required":["repo_url","repo_name","status"],"title":"RepoStatusResponse"},"SearchRequest":{"properties":{"repo_url":{"type":"string","title":"Repo Url"},"query":{"type":"string","title":"Query"},"top_k":{"type":"integer","title":"Top K","default":10}},"type":"object","required":["repo_url","query"],"title":"SearchRequest"},"SearchResult":{"properties":{"chunks":{"items":{"$ref":"#/components/schemas/CodeChunk"},"type":"array","title":"Chunks"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["chunks","total"],"title":"SearchResult"},"SourceReference":{"properties":{"file_path":{"type":"string","title":"File Path"},"start_line":{"type":"integer","title":"Start Line"},"end_line":{"type":"integer","title":"End Line"},"symbol_name":{"type":"string","title":"Symbol Name"},"symbol_type":{"anyOf":[{"$ref":"#/components/schemas/SymbolType"},{"type":"null"}]},"snippet":{"type":"string","title":"Snippet","default":""},"relevance_score":{"type":"number","title":"Relevance Score","default":0.0},"github_url":{"type":"string","title":"Github Url","default":""}},"type":"object","required":["file_path","start_line","end_line","symbol_name"],"title":"SourceReference","description":"Concrete file:line reference in the answer."},"SymbolType":{"type":"string","enum":["function","method","class","module","variable","import","interface","type_alias","enum"],"title":"SymbolType"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}